--- a/Dockerfile
+++ b/Dockerfile
@@ -17,4 +17,5 @@ RUN pip3 install --no-cache-dir -r requirements.txt
 
 EXPOSE 8000
 
-CMD ["gunicorn", "flick.wsgi", "0:8000"]
+# CMD ["gunicorn", "flick.wsgi", "0:8000"]
+CMD ["uvicorn", "flick.asgi:application"]