--- a/requirements.txt
+++ b/requirements.txt
@@ -55,6 +55,7 @@ sqlparse==0.4.1
 toml==0.10.2
 typing-extensions==3.7.4.3
 urllib3>=1.21.1
+uvicorn==0.13.4
 vine>=1.3.0
 virtualenv==20.4.0
 wcwidth==0.2.5