--- a/requirements.txt
+++ b/requirements.txt
@@ -62,4 +62,4 @@ wcwidth==0.2.5
 yarl==1.6.3
 zipp==3.4.0
 tmdbsimple==2.2.15
-qinspect==1.1.0
+django-queryinspect==1.1.0