--- a/src/flick_auth/async_views.py
+++ b/src/flick_auth/async_views.py
@@ -1,0 +1,5 @@
+from django.http import HttpResponse
+
+
+async def index(request):
+    return HttpResponse("Hello, async Django!")