## Overview
* Make notifs faster
```
BEFORE: INFO [SQL] 30 queries (13 duplicates), 3977 ms SQL time, 4638 ms total request time
AFTER: INFO [SQL] 28 queries (12 duplicates), 3260 ms SQL time, 4032 ms total request time
```
* Make suggestions faster
```
BEFORE: INFO [SQL] 20 queries (13 duplicates), 2076 ms SQL time, 2676 ms total request time
AFTER: INFO [SQL] 13 queries (6 duplicates), 1497 ms SQL time, 2075 ms total request time
```
* Add group delete endpoint at `DELETE /api/groups/<id>/`
```
{
    "success": true,
    "data": "Group with id 37 has been deleted."
}
```


## Changes Made
@olivialy524 @ViviYe it's faster / less expensive to use 
```
SomeObject.filter(profile__user=request.user)
``` 
than do 
```
profile = Profile.objects.get(user=request.user)
SomeObject.filter(profile=profile)
```
## Related PRs or Issues
#297 #253